+++ app/docs/TREE_SHAKING.md
@@ -207,6 +207,8 @@ import type { PassportData } from '@selfxyz/common/types/passport';
 - `@selfxyz/common/constants/hashes` - Sample data hashes
 
 #### Utilities
+
+- `@selfxyz/common/utils/attest` - Proving Attestation Utils
 - `@selfxyz/common/utils/hash` - Hash and Poseidon functions
 - `@selfxyz/common/utils/bytes` - Byte manipulation
 - `@selfxyz/common/utils/trees` - SMT and leaf operations
