+++ app/package.json
@@ -75,7 +75,6 @@
     "@noble/hashes": "^1.5.0",
     "@openpassport/zk-kit-lean-imt": "^0.0.6",
     "@openpassport/zk-kit-smt": "^0.0.1",
-    "@peculiar/x509": "^1.12.3",
     "@react-native-async-storage/async-storage": "^2.1.1",
     "@react-native-clipboard/clipboard": "1.13.2",
     "@react-native-community/netinfo": "^11.4.1",
@@ -91,7 +90,6 @@
     "@selfxyz/mobile-sdk-alpha": "workspace:^",
     "@sentry/react": "^9.32.0",
     "@sentry/react-native": "7.0.0-beta.1",
-    "@stablelib/cbor": "^2.0.1",
     "@tamagui/animations-css": "^1.129.3",
     "@tamagui/animations-react-native": "^1.129.3",
     "@tamagui/config": "1.126.14",
