+++ app/src/utils/cryptoLoader.ts
@@ -12,7 +12,6 @@ export const loadCryptoUtils = async () => {
 export const loadProvingUtils = async () => {
   return Promise.all([
     import('@/utils/proving/provingMachine'),
-    import('@/utils/proving/attest'),
     import('@/utils/proving/validateDocument'),
   ]);
 };
