+++ common/tests/genMockPassportData.test.ts
@@ -1,5 +1,4 @@
-import { expect } from 'chai';
-import { describe, it } from 'mocha';
+import { describe, expect, it } from 'vitest';
 
 import { genAndInitMockPassportData } from '../src/utils/passports/genMockPassportData.js';
 import { parsePassportData } from '../src/utils/passports/passport_parsing/parsePassportData.js';
@@ -14,9 +13,7 @@ const testCases = [
   { dgHashAlgo: 'sha1', eContentHashAlgo: 'sha1', sigAlg: 'ecdsa_sha1_secp256r1_256' },
 ];
 
-describe('Mock Passport Data Generator', function () {
-  this.timeout(0);
-
+describe('Mock Passport Data Generator', () => {
   testCases.forEach(({ dgHashAlgo, eContentHashAlgo, sigAlg }) => {
     it(`should generate valid passport data for ${sigAlg}`, () => {
       const passportData = genAndInitMockPassportData(
@@ -27,7 +24,7 @@ describe('Mock Passport Data Generator', function () {
         '000101',
         '300101'
       );
-      expect(verify(passportData, dgHashAlgo, eContentHashAlgo, sigAlg)).to.be.true;
+      expect(verify(passportData, dgHashAlgo, eContentHashAlgo, sigAlg)).toBe(true);
     });
   });
 });
@@ -41,13 +38,13 @@ function verify(
   const passportMetaData = parsePassportData(passportData);
   // console.log('passportMetaData', passportMetaData);
 
-  expect(passportMetaData.dg1HashFunction).to.equal(dgHashAlgo);
-  expect(passportMetaData.eContentHashFunction).to.equal(eContentHashAlgo);
+  expect(passportMetaData.dg1HashFunction).toBe(dgHashAlgo);
+  expect(passportMetaData.eContentHashFunction).toBe(eContentHashAlgo);
 
   // regex to find the signature algorithm (ecdsa or rsa/rsapss) before first underscore
   const signatureAlgorithm = sigAlg.match(/^([^_]+)/)?.[1];
 
-  expect(passportMetaData.signatureAlgorithm).to.equal(signatureAlgorithm);
+  expect(passportMetaData.signatureAlgorithm).toBe(signatureAlgorithm);
 
   return true;
 }
