+++ common/tests/setup.ts
@@ -1,0 +1,8 @@
+// Vitest setup file for common package tests
+import { vi } from 'vitest';
+
+// Global test setup
+global.vi = vi;
+
+// Mock any global objects that might be needed
+global.Buffer = Buffer;
